Implemented forwarding for `NoRetryStrategy`
